+++ android/app/build.gradle
@@ -76,4 +76,6 @@ dependencies {
 
     implementation "androidx.navigation:navigation-fragment-ktx:2.7.5"
     implementation "androidx.navigation:navigation-ui-ktx:2.7.5"
+
+    implementation 'androidx.datastore:datastore-preferences:1.0.0'
 }
