+++ android/app/src/main/AndroidManifest.xml
@@ -47,11 +47,10 @@
             android:name=".LocationPermissionActivity"
             android:exported="true" />
         <activity
-            android:name=".SplashActivity"
+            android:name=".login.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
