--- android/app/src/main/java/com/example/moHaeng/SplashActivity.kt
@@ -1,20 +1,0 @@
-package com.example.moHaeng
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.os.Handler
-import com.example.moHaeng.login.LoginActivity
-
-
-class SplashActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash)
-
-        Handler().postDelayed({
-            startActivity(Intent(this, MainActivity::class.java))
-            finish()
-        },3000)
-    }
-}
