+++ android/app/src/main/java/com/example/moHaeng/login/JwtCheck.kt
@@ -3,6 +3,10 @@ package com.example.moHaeng.login
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
+import android.widget.Toast
+import com.example.moHaeng.BuildConfig
+import com.example.moHaeng.MainActivity
+import com.example.moHaeng.login.LoginActivity.*
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -16,31 +20,10 @@ public class JwtCheck {
         return sharedPreferences.getString("accessToken", null)
     }
 
-//    fun checkJwtFromServer{
-//        //retrofit2를 통해 refresh token을 서버로 보내서 유효한지 확인한다.
-//        //유효하면 true, 아니면 false를 반환한다.
-//        val token = "Bearer " + getJwtToken(this) // 토큰을 가져온 후 "Bearer "를 추가하여 헤더 형식으로 만듭니다
-//
-//        val retrofit = Retrofit.Builder()
-//            .baseUrl("https://your-api-url.com/") // 실제 API URL로 변경
-//            .addConverterFactory(GsonConverterFactory.create())
-//            .build()
-//
-//        val apiService = retrofit.create(LoginActivity.ApiService::class.java)
-//
-//        apiService.validateToken(token).enqueue(object : Callback<Void> {
-//            override fun onResponse(call: Call<Void>, response: Response<Void>): Boolean {
-//                return response.isSuccessful
-//            }
-//
-//            override fun onFailure(call: Call<Void>, t: Throwable) {
-//                // 네트워크 오류 등으로 요청이 실패한 경우
-//                // 에러 처리 로직 추가
-//            }
-//        })
-//
-//
-//    }
+    fun getRefreshToken(context: Context): String? {
+        val sharedPreferences = context.getSharedPreferences("MyPreferences", Context.MODE_PRIVATE)
+        return sharedPreferences.getString("refreshToken", null)
+    }
 
 
     //토큰 확인 실패시 로그인 화면으로 가는 함수
@@ -49,20 +32,19 @@ public class JwtCheck {
         nowActivity.startActivity(intent)
     }
 
+    //토큰 확인 성공시 메인 화면으로 가는 함수
+    fun goToMainActivity(nowActivity: Activity) {
+        val intent = Intent(nowActivity, MainActivity::class.java)
+        nowActivity.startActivity(intent)
+    }
 
 
-
-//    //jwt를 서버로 보내서 유효한지 확인하는 함수
-//    fun checkJwt(jwt: String): Boolean {
-//        return false
-//    }
-//    //jwt를 가져오는 함수
-//
-//    //jwt를 가져오는데 실패하면 null을 반환한다.
-//    //jwt를 가져오는데 성공하면 jwt를 반환한다.
-//
-//
-//    //jwt를 서버로 보내서 유효한지 확인한다.
-//    //유효하면 true, 아니면 false를 반환한다.
+    fun saveJwtToken(context: Context, accessToken: String, refreshToken: String) {
+        val sharedPreferences = context.getSharedPreferences("MyPreferences", Context.MODE_PRIVATE)
+        val editor = sharedPreferences.edit()
+        editor.putString("refreshToken", refreshToken)
+        editor.putString("accessToken", accessToken)
+        editor.apply()
+    }
 
 }
