+++ android/app/src/main/java/com/example/moHaeng/login/SplashActivity.kt
@@ -1,8 +1,11 @@
 package com.example.moHaeng.login
 
+import android.content.ContentValues
 import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
+import android.os.Looper
+import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import com.example.moHaeng.MainActivity
 import com.example.moHaeng.R
@@ -12,24 +15,26 @@ class SplashActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_splash)
 
-//        val jwtToken = JwtCheck().getAccessToken(this)
-//        if (jwtToken != null) {
-//            JwtCheck().goToLoginActivity(this)
-//        }
 
 
-
-        //JwtCheck의 getJwtToken 함수를 호출하여 jwt를 가져온다.
-
-        //val jwt = JwtCheck().
-        //jwt가 null이 아니면 jwt를 서버로 보내서 유효한지 확인한다.
-        //유효하면 MainActivity로 이동한다.
+        val jwtAccessToken = JwtCheck().getAccessToken(this)
 
 
+        //1초간
+        if (jwtAccessToken != null) {
+            JwtCheck().goToMainActivity(this)
+        }
+        else {
+            val jwtRefreshToken = JwtCheck().getRefreshToken(this)
+            if (jwtRefreshToken != null) {
+                JwtCheck().goToMainActivity(this)
+                //todo 서버로 jwtRefreshToken을 보내서 jwtAccessToken을 받아온다.
+            }
+            else {
+                JwtCheck().goToLoginActivity(this)
+            }
+        }
 
-        Handler().postDelayed({
-            startActivity(Intent(this, LoginActivity::class.java))
-            finish()
-        },3000)
+
     }
 }
